backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI/ML APIs - WORKING VERSIONS
openai==0.28.1
elevenlabs==1.1.0

# Audio processing (optional - for local playback)
sounddevice==0.4.6
soundfile==0.12.1

# WebSocket support (for real-time features later)
websockets==12.0

# CORS support (already included with FastAPI)
# HTTP client (for testing)
httpx==0.27.0
pinecone-client==3.0.0